Update diff engine for symlinks stored in the cache.
This patch updates the external diff interface engine for the change
to store the symbolic links in the cache, recently done by Kay
Sievers.

The main thing it does is when comparing with the work tree, it
prepares the counterpart to the blob being compared by doing a
readlink followed by sending that result to a temporary file to
be diffed.

Signed-off-by: Junio C Hamano <junkio@cox.net>