[PATCH] git cvsimport fuzz argument
Add "-z fuzz" argument, passed to cvsps, and clean up argument
processing.  Also, use "cvsps --cvs-direct", which is is somewhat
faster.

Give the user the option of specifying the timestamp fuzz passed to
cvsps.  Looking at the other arguments to it, I can't see anything else
that would be sane to play with.  Also, use --cvs-direct, which speeds
up cvsps for remote repositories and doesn't seem to do anything bad to
local repositories.

Signed-off-by: Tommy McGuire <mcguire@crsr.net>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>