Add commit row context menu and handle left-click on graph lines
Right-click on a context row now brings up a menu allowing the user to
generate a diff between that row and the selected row.  Left-click on
a graph line shows the parent and children connected by the line in
the details pane.  Left-click on a circle in the graph selects that
commit.  Left-click elsewhere in the graph does nothing.

When displaying a diff, the bottom-right file list box behaves
slightly differently now; instead of eliding all other files' diffs,
it now just scrolls the details pane so that the selected file's diff
starts at the top of the pane.

Since the diffs can be rather large, arrange for an update to be done
every 100ms while reading diffs.

Also removed the CVS revision keywords and bumped the version number
to 1.2.