GIT 0.99.4 (release candidate)
This is my first attempt to adjust Debian and RPM to pass
prefix, to prepare the 0.99.4 release.

It updates debian/rules and git-core.spec.in to properly pass
prefix when building binary packages.  It also updates
debian/changelog to make the resulting binary package name
0.99.4; this is not needed on the RPM side (it takes the version
number from the main Makefile).

Signed-off-by: Junio C Hamano <junkio@cox.net>