Fix documentation dependency.
Randal L. Schwartz noticed that 'make install' does not rebuild what
is installed.  Make the 'install' rule depend on 'man'.

I noticed also 'touch' of the source files were used to express include
dependencies, which is a no-no.  Rewrite it to do dependencies properly,
and add missing include dependencies while we are at it.

Signed-off-by: Junio C Hamano <junkio@cox.net>