rev-list --boundary: fix re-injecting boundary commits.
Marco reported that

	$ git rev-list --boundary --topo-order --parents a3aceb7..9451b04

misses these two boundary commits.

        37af753214e39c925ed2d9a8659572804dea24c5
        9e29cef2dddf46e4b235e596fb82736a314835cf

Indeed, we can see that gitk shows these two commits at the
bottom, because the --boundary code failed to output them.

The code did not check to avoid pushing the same uninteresting
commit twice to the result list.  I am not sure why this fixes
the reported problem, but this seems to fix it.

Signed-off-by: Junio C Hamano <junkio@cox.net>