t0000-basic: Add ls-tree recursive test back.
When we updated ls-tree recursive output to omit the tree nodes,
68a391c706375abf8568ebdd9b14e429b6108821 adjusted the old test
so that we do not expect to see trees in its output.  Later,
with 6ed4109d0d085a90010205ddb4693f5a3ddc5db9, we added back the
ability to show both with -t option, but we forgot to update the
test as well.

Signed-off-by: Junio C Hamano <junkio@cox.net>