Fix refs.c;:repack_without_ref() clean-up path
The function repack_without_ref() passes a lock-file structure
on the stack to hold_lock_file_for_update(), which in turn
registers it to be cleaned up via atexit().  This is a big
no-no.

This is the same bug James Bottomley fixed with commit
dc720a7c3acb5a646387ed8aae39a8ba475fb787.

Signed-off-by: Junio C Hamano <junkio@cox.net>