git-gui: Changed term 'check-in' to 'include'.
At least one user was confused by the term 'check-in'; he thought that
clicking that button would commit just that one file, but he wanted to
include all modified files into his next commit.

Since git doesn't really have a check-in concept this really was poor
language to use.  Git does have an update-index concept but that is a
little too low level to show to the user.

So instead we now talk about including files in a commit.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>