git-gui: Fix mouse cursor behavior when in widgets.
The mouse cursor (at least on Windows) seemed to be picking up the
cursor from the sash controls and then never resetting itself back
to the standard text cursor (the I-beam) when it was over a text area
that the user can edit (like the commit buffer) or over a text area
the user can copy from (like the diff viewer).

So now we always set the cursor to left_ptr (which according to the Tk
documentation should be available everywhere) and only for the two text
areas which we use to list file names, as the user clicks in these but
is not permitted to select text.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>