git-gui: Fix list loading corruption introduced by 1644c3ea.
Tcl let me assign two different types of values to the variable $n.
Prior to 1644c3ea $n was the total number of bytes in the string;
but in that commit it also became the current info list for the
current file.  This caused $c < $n to fail as $n was now treated
as 0 and we only loaded the first file in each buffer.

So use a different variable, like $i, instead. 

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>