git-gui: Rename Project menu to Repository.
Since all of the actions in our Project menu actually apply to the
Git concept of a repository, it is a disservice to our users to
call it "project".  This is especially true if Git ever gets any
sort of subproject support, as the term would then most definately
conflict.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>