Revert "git-diff: Introduce --index and deprecate --cached."
This reverts commit b1cce3b9c2474da2d80e0218286234a2b0728d7a.

Although --cached and --index are confusing wording, the use of
word --cached for git-diff is consistent with git-apply.  It means
"work with index without looking at the working tree".

We should probably come up with better wording for --cached, if
somebody wants to deprecate it.  But making --index and --cached
synonyms for diff while leaving them mean different things for
apply is no good.