Revert "git-pull: refuse default merge without branch.*.merge"
This reverts commit 9a3e8fcec565b87726e79935d3ca75397896d1b5.

The logic to decide when to refuse to use the default "first set of
refs fetched" for merge was utterly bogus.

In a repository that happily worked correctly without any of the
per-branch configuration crap did not have (and did not have to
have) any branch.<current>.merge.  With that broken commit, pulling
from origin no longer would work.