git-gui: Use system default labelframe bordering.
In the new branch dialog and delete branch dialog we are using the
system default labelframe border settings (whatever those are) and
they look reasonable on both Windows and Mac OS X.  But for some
unknown reason to me I used a raised border for the options dialog.
It doesn't look consistent anymore, so I'm switching it to the
defaults.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>