Test 'git add' for unmerged entries when core.symlinks=false.
In 33ec31ebf4aa42f0f27d14ff3ecacf7c6b9f1bba git add was fixed if unmerged
entries are in the index and core.filemode=false. core.symlinks=false is
a similar case, which touches the same code path. Here is a test that
makes sure that the symlink property in the index is preserved, too.

Signed-off-by: Johannes Sixt <johannes.sixt@telecom.at>
Signed-off-by: Junio C Hamano <gitster@pobox.com>