Make "git-log --" without paths behave the same as "git-log" without --
"git log" family of commands, even when run from a subdirectory,
do not limit the revision range with the current directory as
the path limiter, but with double-dash without any paths after
it, i.e. "git log --" do so.  It was a mistake to have a
difference between "git log --" and "git log" introduced in
commit de44b9c26b4d9ac1ef81b63a9ad15992ef1f5174 (First cut at
libifying revlist generation).

Signed-off-by: Junio C Hamano <gitster@pobox.com>