gitk: Simplify highlighting interface and combine with Find function
This effectively coaelesces the highlighting function and the search
function.  Instead of separate highlight and find controls, there is
now one set of interface elements that controls both.  The main
selector is a drop-down menu that controls whether commits are
highlighted and searched for on the basis of text in the commit
(i.e. the commit object), files affected by the commit or strings
added/removed by the commit.

The functions to highlight by membership of a view or by ancestor/
descendent relation to the selected commit are gone, as is the
move to next/previous highlighted commit (shift-up/down) function.

Signed-off-by: Paul Mackerras <paulus@samba.org>