git-gui: Ensure copyright message is correctly read as UTF-8
On Windows using the native Tcl/Tk the copyright header is
being read from the script using the system encoding, which
may not be utf-8.  This causes the multi-byte copyright symbol
(which is actually encoded as utf-8) to read as two characters
and not as a proper copyright symbol.  Explicitly asking Tcl
to read this sequence of bytes as utf-8 corrects the issue.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>