Avoid scary errors about tagged trees/blobs during git-fetch
This is the same bug as 70ea0956497a4f9f68d444347481a170ea9c93e9.
The warning "Object $X is a tree, not a commit" is bogus and is
not relevant here.  If its not a commit we just need to make sure
we don't mark it for merge as we fill out FETCH_HEAD.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>