builtin-commit.c: remove useless check added by faulty cut and paste
2a6f7a067dc25c92de1e46e4eb6dac13daf2f9a5 (builtin-commit: fix
partial-commit support) mindlessly cut and pasted from
builtin-ls-files.c, and included a part that was meant to
exclude redundant path after "ls-files --with-tree" overlayed
the HEAD commit on top of the index.  This logic does not apply
to what git-commit does and should not have been copied, even
though it would not hurt.

Signed-off-by: Junio C Hamano <gitster@pobox.com>