gitk: Fix bug causing Tcl error when no commits are selected
Some of the stuff that commit b22d6022a33d2c41115def1af237cfb74f9e34c0
added to drawvisible isn't appropriate to do when we have no commits,
and this was causing a Tcl error if gitk was invoked in such a fashion
that no commits were selected.  This fixes it by bailing out of
drawvisible early if there are no commits displayed.

Bug reported by Johannes Sixt.

Signed-off-by: Paul Mackerras <paulus@samba.org>