filter-branch: handle filenames that need quoting
The command used a very old fashioned construct to extract
filenames out of diff-index and ended up corrupting the output.
We can simply use --name-only and pipe into --stdin mode of
update-index.  It's been like that for the past 2 years or so
since 9b8a991 (update-index: work with c-quoted name).

Signed-off-by: Junio C Hamano <gitster@pobox.com>