git-blame --reverse
This new option allows "git blame" to read an old version of the file, and
up to which commit each line survived (i.e. their children rewrote the
line out of the contents).  The previous revision machinery update to
decorate each commit with its children was leading to this change.  When
the --reverse option is given, we read the old version and pass blame to
the children of the current suspect, instead of the usual order of
starting from the latest and passing blame to parents.

The standard yardstick of "blame" in git.git history is "rev-list.c" which
was refactored heavily in its existence.  For example:

    git blame -C -C -w --reverse 9d6a0499..master -- rev-list.c

begins like this:

8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02  1) #include "cache...
8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02  2) #include "commi...
8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02  3) #include "tree....
8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02  4) #include "blob....
f006ba7c rev-list.c         (JC Hamano   2006-03-01  5) #include "epoch...
8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02  6)
9451b049 rev-list.c         (JC Hamano   2006-02-24  7) #define SEEN
9451b049 rev-list.c         (JC Hamano   2006-02-24  8) #define INTERES...
f006ba7c rev-list.c         (JC Hamano   2006-03-01  9) #define COUNTED...
b8fe5152 rev-list.c         (LTorvalds   2005-07-06 10) #define SHOWN  ...
8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02 11)
8acadf4f builtin-rev-list.c (JC Hamano   2008-04-02 12) static const ch...
0ad47776 rev-list.c         (LTorvalds   2005-07-26 13)    "usage: git-...

This reveals that the original first four lines survived until now in
builtin-rev-list.c , inclusion of "epoch.h" was removed after f006ba7c
while the contents was still in rev-list.c.

This mode probably needs more tweaking so that the commit that removed the
line (i.e. the children of the commits listed in the above sample output)
is shown instead to be useful, but then there is a little matter of which
child of a fork point to show.

For now, you can find the diff that rewrote the fifth line above by doing:

    $ git log --children f006ba7c^..

to find its child, which is 601c5bc (Merge branch 'lt/rev-list' into next,
2006-03-01), and then look at that child with:

    $ git show 601c5bc

Signed-off-by: Junio C Hamano <gitster@pobox.com>