racy-git: an empty blob has a fixed object name
We use size=0 as the magic token to say the entry is known to be racily
clean, but a sequence that does:

 - update the path with a non-empty blob and write the index;
 - update an unrelated path and write the index -- this smudges
   the above entry;
 - truncate the path to size zero.

would make both the size field for the path in the index and the size on
the filesystem zero.  We should not mistake it as a clean index entry.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Junio C Hamano <gitster@pobox.com>