Revert "mailinfo: Remove only one set of square brackets"
This reverts commit 2d0a1cf7834e8d3f2e4fbd1129a79c861a32e364.

Some mailing lists are configured add prefix "[listname] " to all their
messages, and also people hand-edit subject lines, be it an output from
format-patch or a patch generated by some other means.

We cannot stop people from mucking with the subject line, and with the
change, there always will be need for hand editing the subject when that
happens.  People have depended on the leading [bracketed string] removal.